action_recognition_module/object-recognition/requirements_groundedsam_venv.txt
accelerate==0.32.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
antlr4-python3-runtime==4.9.3
black==21.4b2
boto3==1.34.141
botocore==1.34.141
chumpy==0.70
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
colorama==0.4.6
crcmod==1.7
dataclasses-json==0.6.7
datasets==2.21.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@ff53992b1985b63bd3262b5a36167098e3dada02
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
easydict==1.13
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
fairscale==0.4.4
faiss-gpu-cu12==1.8.0.2
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
filterpy==1.4.5
freetype-py==2.4.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gradio==4.37.2
gradio_client==1.0.2
httptools==0.6.1
huggingface-hub==0.23.4
hydra-core==1.3.2
importlib_resources==6.4.0
iopath==0.1.9
jmespath==0.10.0
json-tricks==3.17.3
jsonpatch==1.33
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.116
litellm==1.41.13
marshmallow==3.22.0
mmpose==0.28.0
model-index==0.1.11
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
nltk==3.8.1
omegaconf==2.3.0
onnxruntime==1.18.1
openai==0.28.0
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
orjson==3.10.6
oss2==2.17.0
pathspec==0.12.1
plyfile==1.0.3
portalocker==2.10.0
pyarrow==17.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodome==3.20.0
pydot==2.0.0
pydub==0.25.1
pyglet==1.5.27
PyOpenGL==3.1.0
pyrender==0.1.45
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
regex==2024.5.15
requests==2.32.3
rich==13.4.2
ruff==0.5.1
s3transfer==0.10.2
safetensors==0.4.3
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
smplx==0.1.28
starlette==0.37.2
supervision==0.21.0
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torchgeometry==0.1.2
tqdm==4.66.5
transformers==4.40.1
trimesh==4.4.1
typer==0.12.3
typing-inspect==0.9.0
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
xtcocotools==1.14.3
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.2